move allocator implementation into separate files
